% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatCASiList.R
\name{CreateCASiList}
\alias{CreateCASiList}
\title{Build LASiC object.}
\usage{
CreateCASiList(t0_count, t1_count, t0_label = NULL, nfeature = 2000)
}
\arguments{
\item{t0_count}{Gene expression matrix of the initial data point/t0.}

\item{t1_count}{Gene expression matrix of the following data points/t1.}

\item{t0_label}{Cell labels of the initial data point.}

\item{nfeature}{The number of top highly variable genes. Default to 2000.}
}
\description{
This function allows you to build the LASiC list and all other functions will be applied to this list.
}
