% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossTimeAnnotation.R
\name{CrossTimeAnnotation}
\alias{CrossTimeAnnotation}
\title{Cross-timepoint cell annotation.}
\usage{
CrossTimeAnnotation(CASiList, DrawUMAP = TRUE)
}
\arguments{
\item{CASiList}{The CASi list.}

\item{DrawUMAP}{Draw UMAP plots per time points and cell labels. Default to TRUE.}
}
\description{
This function allows you to annotate t1 cells.
}
