% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectNovelCells.R
\name{DetectNovelCells}
\alias{DetectNovelCells}
\title{Detect novel cells.}
\usage{
DetectNovelCells(CASiList, corr.cutoff = 0.4, p.cutoff = 0.01, DrawUMAP = TRUE)
}
\arguments{
\item{CASiList}{The CASi list.}

\item{corr.cutoff}{The cutoff correlation value to determine novel cells. Default to 0.4. Cells with the correlation larger than 0.4 will not be considered novel cells.}

\item{p.cutoff}{The cutoff p value of t-tests to determine novel cells. Default to 0.01. Cells with the p value smaller than 0.01 will be considered novel cells.}

\item{DrawUMAP}{Draw UMAP plots per cell type-specific correlation and cell labels.}
}
\description{
This function allows you to detect any novel cell type emerged over time.
}
