% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindTemporalDEGs.R
\name{FindTemporalDEGs}
\alias{FindTemporalDEGs}
\title{Identify temporal differentially expressed genes (tEGs).}
\usage{
FindTemporalDEGs(CASiList, metadata, pct.cutoff = 0.3, p.cutoff = 0.05)
}
\arguments{
\item{CASiList}{The CASi list.}

\item{metadata}{The dataframe containing the Group variable and the Time variable.}

\item{pct.cutoff}{The percentage of cells that the gene is expressed. Default to 0.3.}

\item{p.cutoff}{The cutoff p value of t-tests to determine tDEGs. Default to 0.05. Genes with the p value smaller than 0.05 will be considered tDEGs.}
}
\description{
This function allows you to find tDEGs.
}
